Extracted the reusable MarkdownRenderer code from the BasePageRenderer subclass
